The start of a unit test project, initially parts of ajabase.
I have found a simple 1 header unit testing framework that works quite well.
The first parts of ajabase that are getting the unit tests are the new
string helper functions in common/common.h. I have also converted the
old unit test code in AJAPersistence to doctest.

The framework is called doctest and is from: https://github.com/onqtam/doctest.
The version being checked in is from the 1.1.4 tag