Changed AJADebug::Assert to AJADebug::AssertWithMessage
This change was made due to Assert being a macro used in the Mac CMIO
API that was causing name collisions. This new name better describes the
functionality anyway.

Also changed the method to only do anything on debug builds (like the
normal assert call does) and I also changed the throws to calls to the
standard c assert().