Improve the register expert singleton
Before this change I was running into strange crashes when starting
Premiere (which loads multiple AJA plug-ins with multiple threads) under
debug on the Mac. I'm guessing the crash was due to globals not having a
guaranteed initialization order.

Now access to the global single instance of RegisterExpert is
gated by an instance method in CNTV2RegisterExpert which all of its
methods call to access the RegisterExpert singleton.

Ideally the instance method would create an "auto" lock on the stack to
insure only 1 thread is in there at any one time. The reason I did not
put a lock in there is because NTV2 does not have a built-in lock and I
do not want to introduce a dependency on ajabase.

So there still is a little room for improvement but still better than
before.