Changed the way bundling up Qt .so's on Linux for the SDK
Needed to change away from using a qt.conf file since it would break the
local qt built apps that are put in the bin directory (not good for an
SDK).

I ended up placing all the .so's including the qpa platform
library, libqxcb.so, into a qtlibs directory. In the place of the
qt.conf I am now setting the enviornment variable
QT_QPA_PLATFORM_PLUGIN_PATH so the precompiled apps know where to find
the platform library. This environment variable is only set for the Qt
apps we prebuild for the SDK and it is not enabled in the code that goes
out with the SDK, this should allow our prebuilt apps to co-exist with
customer built apps while all in the same bin directory.