Ntv2 diagnostics now compile in the 14.x branch.
The diagnostics were last build with the 12.4 or 12.5 branch, skipped
right over 13. A lot of old code was that was stripped out of 13 is
needed for this app, rather than put it back into the SDK I have put it
into the app folder directly. Besides finding all the missing files the
big issue was bringing it up to date with the latest SDK so it can build
with NTV2_DEPRECATE. The third axis of the overhaul was getting it off
of VisualStudio 2008 and on to VisualStudio 2013.