Moved the CNTV2SupportLogger out to its own class in ajantv2.
The idea behind CNTV2SupportLogger is to get the log generation code out
of the watcher and into the SDK so OEM's can also use it within their
applications if desired.

The first client of this class is the new "supportlog" command line app
being checked in with this commit. Currently we have a rebranded
ntv2watcher binary named supportlog in the SDK which we give to OEM's
but since it is a Qt app it will not run on a headless linux server,
this new command line supportlog will. The new supportlog is a small
main wrapped around AJASystemInfo (from ajabase) and CNTV2SupportLogger
(from ajantv2).

I would like to use CNTV2SupportLogger as the foundation for the support
log functionality in ntv2watcher in the near future.