Changes to AJASystemInfo so can better control when a Rescan() is done
Added a new enum type AJASystemInfoSections, the enums of this type are
bitmasks so groups can be selectively used in the contructor and
Rescan(). Some of the info information is a bit expensive to gather,
like the GPU information on Mac and Linux. Now the user can choose some,
none or all of the information to rescan at instatiation time or when
doing a Rescan().

Using this new functionallity I have updated AJAPersistence and the
ControlPanel infopane to only scan for the AJA_SystemInfoSection_Path
section, greatly reducing the time spent scaning on the Mac (GPU
section).

I've implemented the change in such a way than any existing uses of
AJASystemInfo will not require any changes, new functionality is opt-in.