Modification to NTV2GetBitfileName() so can get OEM bitfile name on Windows
Before this change this function would return the OEM bitfile names on
Mac and Linux and the cleaned up retail names on Windows. There was no
way to get the OEM name on Windows.

Now there is a useOemNameOnWindows parameter which defaults to false, if
set to true the OEM bitfile name will be returned on Windows. Existing
calls to this function will not need to be changed, it is opt-in to get
the OEM names.