Added a 4th way that AJASystemInfo on Windows gets the current username
HBO is running into issues when setting up machines remotely using
Control Panel which uses AJASystemInfo to get the path to the
persistence file. Added a 4th way that will be tried first by calling
SHGetFolderPathA() to get the users home folder/name.

In my testing this new way works as well as the previous method being
used first, because of HBO's setup we will have to build and throw over
the wall for them to test.