Drivers were not checking LTC status when AUTOCIRCULATE_WITH_LTC was set
Before this change AutoCirculate could be inited with LTC but there was
no way of reading back that value from user space.

Fix can be verified by seeing that the LTC state changes from NO to YES
in the AutoCirculate pane of the watcher.

The CNTV2SupportLogger class also needed an update to output the LTC
status.