Adding safe SDK calls to zero out the ANC regions of frames
I was going down the road of adding this functionality to Streams so the
ANC regions will be cleared out after Auto Circulate is initialized for
playback (so CC data will be cleared when switching between clips in
Premiere). After talking with bills@ we decided the functionality could
be useful to others in the SDK.

The main method is CNTV2Card::DMAClearAncRegion() which can be used to
make sure that the various ANC regions are zeroed out.