Fix for bugs# 17034 & 17048 crashing in Transmit plug-in
The bug was caused by -- being used on an const_iterator that happened to be at
map.begin(), -- is undefined behavior in this case.

Added a check for this in CNTV2Card::GetAncRegionOffsetAndSize().

I was only ever able to reproduce this bug when running release builds
on the Mac, debug builds worked fine.