Adding an optional argument of bQuiet to CNTV2Card::ProgramMainFlash()
The bQuiet argument defaults to false, so the existing default behavior
will not change.

The reason this argument was added was so that when
ntv2firmwareinstallerthread (which uses ProgramMainFlash()) is used from
ntv2firmwareinstaller the command line -q (--quiet) argument will be honored.

I have a need for this flag to be honored in the qa automation testing because
all the terminal output is logged and it gets flooded with all the flash
status updates.