Add new flags for Makefile based on ticket# 4137
An OEM asked for a way to not build the demo applications, so now added
support for the environment variable AJA_NO_DEMOS, if that is set the
demo apps will not be built.

While I was in there I added AJA_NO_GL, which will slip the builing of demo
apps that need OpenGL, currently only the nvidia dvplowlatencydemo. I
added this one since sometimes getting OpenGL setup on Linux is a pain
depending on the distro.