Fix bug# 17401, ControlPanel only finding firmware for user who installed
My first thought was that the location where the firmware was put
by the installer was the issue, so I changed the firmware install
location to: c:\ProgramData\AJA\ntv2\Firmware

But that was not the problem, after further testing the problem was due
to the installer writing the registry key for the firmware path into
HKEY_CURRENT_USER, which will only be visible by the user who installs
our retail software. In the SDK within ntv2utils is a function to get
the firmware path which reads this registry key, which of course is how
Control Panel knows where to find the firmware. <Head Slap>

So the solution is to write the registry key under HKEY_LOCAL_MACHINE
which all users can read, also updated the code in ntv2utils to look for
the key in HKEY_LOCAL_MACHINE.

So the move of the firmware to c:\ProgramData\AJA\ntv2\Firmware was not
really needed but I do think it is a better location to put the
firmware, so I am going to retain that change.