A complete fix for bug# 17656
Turns out using read()/write() with a file descriptor does not always
work on Mac.

Stan stumbled on this and did a partial fix with commit# 46855824,
but that then broke the unit tests becasue the file descriptor was
not being saved off on Mac. So that gave me the idea to put back the
code like I had it with the file descriptor saved off (fixes Mac unit
tests) and then change the AJAFileIO::Read() to use fread() and
AJAFileIO::Write() to use fwrite(). Which then broke the unit tests on
Mac and Linux because the parameters to fread() and fwrite() were in the
wrong order.

Long story short, this fixes the issues seen on Mac with Control Room
and the unit tests pass.