Cleanup of Posix path of AJAFileIO
We don't need to cache the filedescriptor in the AJAFileIO class since
we always have the FILE* pointer which contains the filedescriptor. So I
changed the code to call fileno() on FILE* where needed instead of
caching the value in Open(). The code is now more self consistent and
easier to follow.