Added an extra parameter to AJAWavWriter ctor so can write float audio
The current implementation works fine writing out float audio data, just
need to set a flag so that the WAV header gets a 3 instead of a 1 in
one byte. The flag in the constructor defaults to false, so PCM will be
used just as it always has.

This flag was added so I could write out debugging audio files for
ndibridge while working on the ffmpeg based resampling code.