m1: Fix alignment of structure used by AJADebug system
The operations in AJAAtomic were crashing whenever a message was being
sent with AJADebug, turned out to have nothing to do with the atomic
operations but with the alignment of the structure those operations
were working on. In the past the structures had a 1 byte alignment which
works fine on x86_64 hardware. After some internet searching I thought
an alignment of 4 would do the trick but it still crashed. I found some hand
wavy apple docs that mentioned 16 byte alignment for arm stuff so I gave
that a try and it works.