m1: Fix allocation of shared memory on M1 mac
In AJAMemory::AllocateShared() was getting a failure on the call to
mmap() on the M1 mac with native arm64 binaries. I fixed by changing
the string passed to shm_open() to not start with a "/" like linux does.

I tested this fix with both x86_64 and arm64 binaries on the M1, I
even tried a x86_64 version of ajalogger getting log messages from
an arm64 version of Control Room.

NOTE: This change will break compatibility on the Mac for using a newer
ajalogger with older binaries and vice versa.