cmake: Combine all the Qt find_package commands into one line
According to a blog entry[0] on kitware.com (creators of CMake) only one
find_package() should be used with multiple arguments. The reasoning for
this is that if a component is not in one version of Qt but in
another you could get a situation where the needed components are
pulled from more than 1 version of Qt, which would not be a good thing.

The documentation[1] for cmake-qt(7) now uses this method.

[0] - https://blog.kitware.com/cmake-finding-qt5-the-right-way/
[1] - https://cmake.org/cmake/help/v3.10/manual/cmake-qt.7.html