Change offset, conflicts with 12Bit LUT
Update per Jeff L:  “The multi-viewer registers were put at offset byte address 0xf000, register 0x3c00 ( 15,360 ) .  You can move the offset for the multi-view registers to 0x10000 and that should fix the issue.  Only one customer has this firmware and Scott and I can work this out when he returns.  The code for the multi-view registers is probably in the driver.“