persistence: Bump the number of retries and increase time between retries
While looking into issue ntv2/ntv2#844 for Linux I was noticing in the
logs some DB failures due to the DB being locked.

Once I increased the number of retries from 4 to 6 and the time between
retries from 2000 to 2500 microseconds I am not longer seeing these failures
in the logs.