ajantv2: Fix a regression in AutoCirculateInitForInput/Output
The regression was caused by:
ef91732d7118cc793f8274d2ffc382808203e4a5
7b727e24a8b74ed2e2782f8020df85c7291a7fe2

The above commits caused the passing of NTV2_AUDIOSYSTEM_INVALID for the
audioSystem to no longer work, the passing of NTV2_AUDIOSYSTEM_INVALID
is used as a flag to init AutoCirculate without audio.

The fix is to only do the extra audioSystem checks if the passed in
audioSystem is not NTV2_AUDIOSYSTEM_INVALID.

This was the cause of issue ntv2/ntv2#1042, in After Effects we do not
use audio with AutoCirculate so the initialization was failing, which
results in no video output.