Allow overriding Linux driver kernel version in Makefile.
Makes it possible to build the driver in a Docker container.

Containers do not have a typical kernel, and uname will return

a value representing the host system container environment's kernel.

Provided the container has headers installed for the desired kernel,

we can specify the kernel version manually in the Makefile and compile

the NTV2 driver on the container for that kernel version.