Fix COM initialization and cleanup in Windows getOSName_WMI call.
CoUninitialize should only be called after a successful call to CoInitialize.

Re-worked cleanup of COM objects to handle everything in one place.

Handle COM Security and Blanket settings in scope instead of using global calls

If any additional WMI code is added to ajabase in the future, use this function as a starting point and refactor into a more general WMI query class