fix: Don't use CMAKE_BUILD_TYPE
CMAKE_BUILD_TYPE usefullness depends on the generator being used.
Sometimes there is a single configure for Release and Debug and that
doesn't work this way.

There is an option to do config type based defines using generator
expressions. Rewrote this part to include this for the NDEBUG define.

NDEBUG was causing our software to be partially built with NDEBUG in
debug builds causing issues with header libraries we use.