NTV2 Nub, Windows DLL Dependency FIx
In release builds, LoadLibrayExA() was reporting that the specified module was not found.
A WinDbg session revealed that it was other dependent DLLs such as NETAPI32.dll which were not found.
Changing the flag argument to LOAD_LIBRARY_SEARCH_DEFAULT_DIRS.
This includes the same search locations as the previously used LOAD_LIBRARY_SEARCH_USER_DIRS,
but proceeds to search the other "default" DLL search paths also.  Such as C:\WIndows\System32.

Issue #372