NTV2NubAccess : some constants to lower case
Making constants that might be part of a URL String Query lower case.
ParseQueryParams makes all of it's keys lower case as they are added to mParams,
So these constants would not be usable to match query parameters unless they are lower case also.

A specific fix that this enables is to use the deviceindex parameter in the ntv2 nub client.

Issue #372