Nub updates
Defines separate Enocde and Decode functions for the Client and Server.
This allows for optimization... not transmitting unnecessary data depending on reads or writes
particularly while using NTV2Buffer.

Also defined NTV2Buffer encode & decode differently than the rest, allowing it's
buffer content processing to be skipped in the specific cases where it is not needed.

Implemented optimizations for NTV2GetRegisters.

Bug fix for NTV2_HEADER, now skips the decode of fSizeInBytes.  This value can vary
by OS and thus should stay native to the OS that the structures are created in.