ntv2utils : NTV2GetFirmwareFolderPath() return default value for Windows
NTV2GetFirmwareFolderPath() is used to find the plugins folder which is derived from it.
If the retail installer has not been run, the firmware path will not be in the windows registry, and this returned an empty string.
Now returning a default path in this scenario so plugins can be loaded,
such as a nub remote client launched by ntv2watcher or anything else.

Issue #183