NTV2message driver change: Do not fail on non-zero fOperation
fOperation is now borrowed in ntv2 nub to send a connection ID between the client and server.
It is convenient to allow it to stay there rather than clear it before it reaches the driver.
The driver still does not do anything with fOperation, but the change allows something non-zero to exist there.