cmake: Fix missing FileDescription field in Windows resource file
The aja_versionize_windows CMake helper function passes a description
string via the AJA_PRODUCT_DESC CMake variable, which was intended to
be used to populate the FileDescription field in the win-ver.rc.in
template file. This commit fixes the missing template variable expansion
in the win-ver.rc.in CMake template file.